requirements.txt
infinity-sdk==0.5.0.dev6
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
certifi>=2023.7.22